Changed file name from main(4) to api, and merged api changes made from the dashboard branch
